+++ snyk/dependency-sync.py
@@ -7,8 +7,7 @@ def sync():
     pyproject = tomlkit.loads(Path("pyproject.toml").read_text())
     dependencies = pyproject["project"]["dependencies"]
     dev_dependencies = pyproject["project"]["optional-dependencies"]["development"]
-    with open(".snyk/requirements.txt", "w") as req:
-        req.write("# Auto generated\n")
+    with open("snyk/requirements.txt", "w") as req:
         for dep in dependencies:
             req.write(f"{dep}\n")
         for dep in dev_dependencies: